pyproject.toml
[tool.poetry]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = ["Kauet Dias <kauetdiaz@gmail.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-jwt-extended = "^4.7.1"
sqlalchemy = "^2.0.38"
werkzeug = "^3.1.3"
click = "^8.1.8"
flask-migrate = "^4.1.0"
psycopg2-binary = "*"
gunicorn = "*"
flask-bcrypt = "*"
flask-marshmallow = "^1.3.0"
marshmallow-sqlalchemy = "*"
apispec-webframeworks = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
